represent reals as double

+++ src/pdf/types.h
@@ -1,6 +1,6 @@
 #ifndef PDF_TYPES_H_
 #define PDF_TYPES_H_
-typedef float    PDF_REAL;
+typedef double   PDF_REAL;
 typedef int32_t  PDF_INT;
 typedef int32_t  PDF_BOOL;
 typedef uint8_t* PDF_STRING;
